# Add Quadro M5000 graphics card to list