Got this output from the v27 benchmark initially;
```
2020-06-09T14:00:38.426 INFO bellperson::groth16::prover > Bellperson 0.8.0 is being used!
2020-06-09T14:00:38.901 INFO bellperson::gpu::locks > GPU is available for FFT!
2020-06-09T14:00:39.121 INFO bellperson::gpu::fft > FFT: 1 working device(s) selected.
2020-06-09T14:00:39.122 INFO bellperson::gpu::fft > FFT: Device 0: Quadro M5000
2020-06-09T14:00:39.122 INFO bellperson::domain > GPU FFT kernel instantiated!
2020-06-09T14:00:39.488 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:39.877 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:40.329 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:40.670 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:40.976 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:41.340 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:41.341 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:41.681 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:41.915 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:42.286 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:42.287 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:42.644 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:42.902 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:43.260 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:43.260 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:00:43.610 WARN bellperson::multiexp > Cannot instantiate GPU Multiexp kernel! Error: GPUError: No working GPUs found!
2020-06-09T14:00:44.234 INFO storage_proofs_core::compound_proof > snark_proof:finish
2020-06-09T14:00:44.234 INFO filecoin_proofs::api::post > generate_window_post:finish
```

After running with `BELLMAN_CUSTOM_GPU="Quadro M5000:2048"` the output became;
```
2020-06-09T14:11:48.428 INFO bellperson::groth16::prover > Bellperson 0.8.0 is being used!
2020-06-09T14:11:54.102 INFO bellperson::gpu::locks > GPU is available for FFT!
2020-06-09T14:11:54.336 INFO bellperson::gpu::fft > FFT: 1 working device(s) selected.
2020-06-09T14:11:54.336 INFO bellperson::gpu::fft > FFT: Device 0: Quadro M5000
2020-06-09T14:11:54.336 INFO bellperson::domain > GPU FFT kernel instantiated!
2020-06-09T14:12:00.094 INFO bellperson::gpu::locks > GPU is available for Multiexp!
2020-06-09T14:12:00.319 INFO bellperson::gpu::utils > Adding "Quadro M5000" to GPU list with 2048 CUDA cores.
2020-06-09T14:12:00.320 INFO bellperson::gpu::multiexp > Multiexp: 1 working device(s) selected. (CPU utilization: 0)
2020-06-09T14:12:00.320 INFO bellperson::gpu::multiexp > Multiexp: Device 0: Quadro M5000 (Chunk-size: 3524235)
2020-06-09T14:12:00.320 INFO bellperson::multiexp > GPU Multiexp kernel instantiated!
2020-06-09T14:12:34.281 INFO filecoin_proofs::api::seal > snark_proof:finish
```