@RobQuistNL 2020-06-11
Benchmark:

CPU: 2 x E5-2680 v3 (12 core 24 thread each)
GPU: Quadro M5000
RAM: 64GB Registered DDR4 @2133 mhz
STORAGE: 500GB S-ATA SSD

**With GPU**
```
BELLMAN_CUSTOM_GPU="Quadro M5000:2048" FIL_PROOFS_USE_GPU_COLUMN_BUILDER=1 RUST_LOG=info FIL_PROOFS_USE_GPU_TREE_BUILDER=1 ./bench sealing --sector-size=512M
results (v27) (536870912)
seal: addPiece: 16.562566267s (30.9 MiB/s)
seal: preCommit phase 1: 6m57.033137125s (1.23 MiB/s)
seal: preCommit phase 2: 54.370806101s (9.42 MiB/s)
seal: commit phase 1: 162.536438ms (3.08 GiB/s)
seal: commit phase 2: 46.170428749s (11.1 MiB/s)
seal: verify: 23.591471ms
unseal: 6m58.577308831s  (1.22 MiB/s)
generate candidates: 291.931µs (1.67 TiB/s)
compute winnnig post proof (cold): 10.867271155s
compute winnnig post proof (hot): 10.732077394s
verify winnnig post proof (cold): 74.654418ms
verify winnnig post proof (hot): 30.783859ms
compute window post proof (cold): 3.167591462s
compute window post proof (hot): 3.167658088s
verify window post proof (cold): 35.759758ms
verify window post proof (hot): 25.214814ms
```

**Without GPU:**  (gave the errors as seen above)
```
FIL_PROOFS_USE_GPU_COLUMN_BUILDER=1 RUST_LOG=info FIL_PROOFS_USE_GPU_TREE_BUILDER=1 ./bench sealing --sector-size=512M
results (v27) (536870912)
seal: addPiece: 17.014401617s (30.1 MiB/s)
seal: preCommit phase 1: 6m59.27734897s (1.22 MiB/s)
seal: preCommit phase 2: 1m28.687148104s (5.77 MiB/s)
seal: commit phase 1: 151.870984ms (3.29 GiB/s)
seal: commit phase 2: 42.264983024s (12.1 MiB/s)
seal: verify: 23.749382ms
unseal: 6m58.435648385s  (1.22 MiB/s)
generate candidates: 263.273µs (1.85 TiB/s)
compute winnnig post proof (cold): 14.160977751s
compute winnnig post proof (hot): 14.204081158s
verify winnnig post proof (cold): 82.262937ms
verify winnnig post proof (hot): 19.705165ms
compute window post proof (cold): 5.955235814s
compute window post proof (hot): 5.934802739s
verify window post proof (cold): 31.061805ms
verify window post proof (hot): 24.664671ms
```