@RobQuistNL 2020-06-12
I see its stating the following;
```
Diff in /home/circleci/gpuci/src/gpu/utils.rs at line 31:
+
-
```

But I have no idea what that means :sweat_smile: 

Wondering if its just a newline thing... Removed one newline :)