@lucas556 2020-08-09
Mine also had this problem, causing delays.